removed hybridizegridpass (poor quality) and smtPrinter (uses Z3 but no z3 library in repository)
